+++ src/providers/ga/ga.js
@@ -57,6 +57,12 @@ analytics.addProvider('Google Analytics', {
     track : function (event, properties) {
         properties || (properties = {});
 
+        var value;
+
+        // Since value is a common property name, ensure it is a number
+        if (analytics.utils.isNumber(properties.value))
+            value = properties.value;
+
         // Try to check for a `category` and `label`. A `category` is required,
         // so if it's not there we use `'All'` as a default. We can safely push
         // undefined if the special properties don't exist.
@@ -64,7 +70,9 @@ analytics.addProvider('Google Analytics', {
             '_trackEvent',
             properties.category || 'All',
             event,
-            properties.label
+            properties.label,
+            value,
+            properties.noninteraction
         ]);
     },